Adding functionality for image upload and exporting orders as PDF

+++ controllers/admin.js
@@ -1,6 +1,7 @@
 const { validationResult } = require('express-validator/check');
 
 const error = require('../utils/error');
+const fileHelper = require('../utils/file');
 
 const Product = require('../models/product');
 
@@ -17,9 +18,27 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
+
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/edit-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description
+      },
+      errorMessage: 'Attached file is not an image',
+      validationErrors: []
+    });
+  }
+
+  const imageUrl = image.path;
 
   const errors = validationResult(req);
 
@@ -31,7 +50,6 @@ exports.postAddProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: title,
-        imageUrl: imageUrl,
         price: price,
         description: description
       },
@@ -93,7 +111,7 @@ exports.postEditProduct = (req, res, next) => {
   const title = req.body.title;
   const price = req.body.price;
   const description = req.body.description;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
 
   const errors = validationResult(req);
 
@@ -105,7 +123,6 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: title,
-        imageUrl: imageUrl,
         price: price,
         description: description,
         _id: prodId
@@ -117,14 +134,18 @@ exports.postEditProduct = (req, res, next) => {
 
   Product.findById(prodId)
     .then((product) => {
-      if (product.userId.toString() !== req.userId.toString()) {
+      if (product.userId.toString() !== req.user._id.toString()) {
         return res.redirect('/');
       }
 
       product.title = title;
       product.price = price;
       product.description = description;
-      product.imageUrl = imageUrl;
+
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
 
       return product.save().then((result) => {
         res.redirect('/admin/products');
@@ -138,9 +159,20 @@ exports.postEditProduct = (req, res, next) => {
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
 
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
-    .then((result) => {
-      res.redirect('/admin/products');
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(err);
+      }
+
+      Product.deleteOne({ _id: prodId, userId: req.user._id })
+        .then((result) => {
+          fileHelper.deleteFile(product.imageUrl);
+          res.redirect('/admin/products');
+        })
+        .catch((err) => {
+          return error.throwError(err, next);
+        });
     })
     .catch((err) => {
       return error.throwError(err, next);
